+++ examples/parse_hpac.py
@@ -8,7 +8,7 @@ with nested arrays for atmospheric area and ionosphere data blocks:
 
 Usage:
 
-python3 parse_hpac.py infile="hpac.log" key="930d847b779b126863c8b3b2766ae7cc", basedate=451169309
+python3 parse_hpac.py filein="hpac.log" key="930d847b779b126863c8b3b2766ae7cc", basedate=451169309
 
 Basedate must be in 32-bit gnssTimeTag integer format - use date2timetag() to convert datetime.